log at debug level when waiting for the src dir
I've observed that it typically takes about 1m for the src dir to become available when
using git-sync to clone our kubernetes manifests. With the log messages as they are now,
this results in ~60 messages logged at WARN level before the directory is found.

This seems wrong and could potentially be misleading, given that this is essentially the
application behaving as expected. So,  I've changed the log level of that message to DEBUG
and also added a couple of INFO level messages to communicate what is happening.