prevent panic when logging errors produced by WaitForDir
We were passing the error returned by WaitForDir as the second argument to
log.Logger.Error, which is expected to be a string.

This was producing a panic:
```
panic: interface conversion: interface {} is *errors.errorString, not string
```

I've amended the arguments and also tweaked the format of the logs a bit.