Restore output on errors (#133)
* Restore output on errors

It's possible for the error output of kubectl to leak secrets. Initially
we addressed this by dropping all error output. This commit attempts to
readdress this in a more nuanced way.

For kustomize we can extract secrets from the YAML, apply that
separately and then drop only that output on error. This gives us 100%
confidence that we won't leak secrets.

For the other mode, where we apply manifests with `kubectl apply -f
<path>`, this adds filtering that will squash output that contains
the terms `Secret` or `base64`. This addresses all instances of this
leak that we're aware of but could potentially leak exceptions.

Errors will continue to be omitted from logs, which covers us against
one of these theorectically exceptions being persisted through log
aggregation.

* golint

* Extract err output terms into a list

This surfaces the criteria more obviously and makes it simpler to update
the terms.

* Reduce verbosity and repetition in kubectl apply methods

By extracting common logic into shared methods.