Prune unchanged dirs from the apply list in Runner
This should help speed things up since we're no longer applying
directories which have had no changes in git since kube-applier's last
run.

However, since git-sync might update the underlying repository while
kube-applier is running, we might have false positives; assuming that:
- lastRunHash is A,
- the current HEAD is at B when the run starts,
- git-sync updates the repo while kube-applier is still applying, with
  the new HEAD at C,
the value of lastRunHash at the end of the run is B.

In the next run that is queued, the Runner will check for changes since
commit B, ensuring that no changes are missed, however, some of the
directories that have changes in both checks (A to B and B to C) might
have been applied already from the first run.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>