Check last applied hash per directory
The last applied hash is stored per directory and is only updated on a
successful apply.

This ensures that if an apply fails for a certain directory, it will not
be ignored on the next run (either a forced run or a new commit that
introduces no changes to that directory).

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>