Use relative paths
Use relative paths since the absolute path depends on how kube-applier
is configured and deployed. In turn, this allows us to easily hide that
information from the ui, which is not useful and also paves the way to
using a global temporary copy of the repositor for apply runs, in order
to avoid races with git-sync.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>