Add support for --server-side (#151)
Create an annotation that can toggle server-side apply on. Defaults to
false, at least while this feature is still beta.

This increases the number of positional arguments to the point that I
think it's worth creating a struct for passing optional flags to the
apply command. So I've done that.

This also moves the inference of whether to do a kustomize apply into
the kubectl package.