Set --force-conflicts with --server-side (#152)
A conflict occurs when a server-side apply updates a field which is
managed or co-managed by another 'field manager'. It causes kubectl to exit
with status 1.

This could be useful for identifying instances where kube-applier is
jostling with some other process for management of a resource but mostly
I imagine this will be cases where someone has performed a manual client-side
apply.

Kube-applier should be applying the source of truth for the set of
manifests, so it seems right to me that it forcibly coerce things to
match its perspective.