Custom equality check for ApplicationStatusRunInfo
metav1.Time uses RFC3339 for serialisation which causes the value to
lose the nanosecond values and therefore makes reflect.DeepEqual as well
as time.Equal to be false when comparing the two values.

The conversion and loss of nanosecond accuracy seems to happen when
updating the Status of the Application.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>