Merge BatchApplier in Runner
The functionality of BatchApplier is no longer so discrete from that of
the Runner. Essentially, the two perform the task of processing
Applications and applying them but the distinction is not very clear.

Additionally, BatchApplier uses a number of common structs and clients with
the Runner and it will need to share more moving forward.

For these reasons, the two are being merged which will reduce code
complexity overall.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>