Revisit Runner
The Runner is no longer performing batch apply runs, instead it applies
each namespace independently. Starting the Runner returns a channel
which can be used to send Requests to it.

A number of features that no longer make a lot of sense and would need a
redesign too (such as metrics and the concept of a "Last Run") have been
left intact simply to keep the changes in this commit smaller.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>