Do not use --depth for creating repository clones
This does not result in savings and in some cases where the depth is
quite large, it makes the cloning significantly slower compared to a
full clone.

Additionally, by omitting the file:// scheme for the source, git will
simply use hardlinks to copy the repository to the new location which
is, as expected, extremely fast and faster than any other option.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>