Rewrite the Dockerfile
This makes it easier to build images locally by taking advantage of
layer caching. A lot of binaries need to be pulled to test and build the
go binary and the resulting image. The changes in this commit should
reduce data transfer and build times for iterating on feature branches.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>