Remove Prometheus struct from metrics package
We do not use multiple Prometheus structs but instead create and
initialise a single one which is then shared with other structs. It does
not hold any state and its attributes are effectively treated as global.
Therefore, there is no good reason for having the necessary methods as
part of a struct. The package can export them directly and it simplifies
the code, we can simply import and use the metrics package directly,
instead of having to pass down struct references.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>