Check the entire Application for changes
This prevents a situation where the status of an Application has been
updated during a run (eg.: ForcedRun) and every ScheduledRun afterwards
fails to update the Status appropriately because the object has changed
in the apiserver. This race can still occur but it should correct itself
during the next ScheduledRun.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>