Increase test wait duration
Old duration would rarely result in a test failures since the run had
not enough time to queue. The new duration seems to have eliminated the
issue although, ideally, we should use Eventually() for this check.
However, since that would require changing the way the queue is consumed
in tests, this mitigation might prove just enough.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>