Remove kubebuilder dependency
kubebuilder is not an actual dependency and is only installed in order
to pull the binary assets required for the envtest package. This change
pulls a controller-runtime script that helps with installing these
assets, independently of kubebuilder. This means that we don't depend on
the latter to release a new version which has pulled updates from
upstream.

This is apparently the recommended method (also how kubebuilder sets up
new projects), see:
 - https://github.com/kubernetes-sigs/controller-runtime/pull/1092
 - https://github.com/kubernetes-sigs/kubebuilder/blob/master/testdata/project-v3/Makefile#L17-L21

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>