Remove kube-applier ClusterRole
This role is not used. Namespaces that are served by kube-applier need
to bind the admin role to the kube-applier ServiceAccount since it needs
to be able to operate on all resources.

Additionally, it introduces a ClusteRole is the kube-applier namespace
which can complicate things because only a single namespace can be
managing cluster resources.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>