Add autoApply attribute in the Waybill spec
When a Waybill's autoApply attribute is set to false, any type of run
other than forced (ie. manually triggered by the user) is ignored.

The attribute is a *bool because we want it to be optional and default
to true. With a bool type:
 - if the json tag has omitempty, it's not possible to set it to false,
   since that is the empty value and it causes it to default back to true
 - otherwise, it is set to false unless explicitly specified, which is
   something we would like to avoid

The attribute is accessed with `pointer.BoolPtrDerefOr()` (and default
value of true) although at any point in the code the attribute _should_
have a value since it is coming from the apiserver. Still, to avoid any
unforseen panics, we can ensure that it is accessed safely at all times.

Additionally, we export this new boolean as a metric, to allow for
alerts on Waybills with autoApply disabled.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>