Address a test data race
Writing the gitLastQueuedHash attribute from the tests (which is
essentially a hack as explained in the comment) produces a data race.

We can use the waybillsMutex of the Scheduler to lock a bit earlier,
covering both the Waybills map and the gitLastQueuedHash string.

Although this is overreaching a bit (the scope of the mutex is only the
map), it's a small change that addresses the data race appropriately.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>