Capture run setup errors in kubernetes events
Errors during the setup steps of a run request are currently logged by
kube-applier but not presented to the user. The implementation of
setRequestFailure() is incomplete and so they are never presented to the
user.

These errors that do not originate from kubectl better fit in the
concept of a kubernetes event. Instead of updating the status
subresource of the Waybill, we can instead emit events. This will
maintain the information of the actual last run in the status of the
Waybill but still surface the issue to the user.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>