Set default for Waybill spec
Since all attributes are optional, we should allow omitting the Waybill
spec. Currently this means that default values for the attributes do not
apply, which causes kube-applier to panic.

Making the filed required means we have to define `spec: {}` in the
Waybill definition and it doesn't look very clean. Instead, we can
define a default value for it.

However, the controller-tools generator currently seems to ignore
an emtpy object `{}` as a default value. We could set this manually but
it becomes error prone if we have to manually fix every time we generate
the CRDs. Instead we can simply define a single attribute in this
default object and `autoApply` was chosen because it is a default that
is very unlikely to change.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>