Restrict kube-applier ClusterRole access to Secrets
Since the strongbox keyring secret will now be fetched using the
delegate token, we can restrict its access to Secrets. For Waybills that
use a non-default secret name for the delegate token, a separate
Rolebinding is required that allows the kube-applier server
ServiceAccount to access the secret. This can be set up in the
client namespace.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>