Move gitPollingLoop in a separate method
Fixes a context leak where cancel is not called at the end of the case.
Instead of simply calling cancel there, however, moving the logic into a
discrete method simplifies the whole approach by allowing us to use
defer. This should make it a bit harder to leak the context in the
future and makes the code easier to read.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>