Extract applyWorker logic into processRequest
Similar to 2f79b98af3b11c490c1c46f233ff675f41e416da, this simplifies
cleanup for the context. Although there is no leak reported by go vet,
the contexts do not actually clean up properly until the worker exits.
However, the context is only relevant for the individual run and should
cleanup when it finishes.

Additionally, this simplifies cleanup of temporary directories which is
something that's regressed at least a couple times in the past.

Using a separate method and relying on defer makes this less error
prone.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>