Rework git.Repository.StartSync
Fixes a race with a nil channel that usually causes the WaitForSync()
method to timeout. The new implementation is simpler, StartSync() will
synchronously perform the initial clone and then spawn the sync loop in
a goroutine internally. This matches the behaviour of the Runner and
Scheduler Start() methods which spawn their respective goroutines
internally.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>