Change GitPollInterval to GitPollWait
This change is to avoid an error that regularly occurs: the Scheduler
needs to check the repository HEAD hash in order to detect changes,
however the Repository also needs to regularly check for updates in the
remote. Both of these operations need to lock the internal mutex which
leads to the latter failing due to the relatively short timeout.

With this change, instead of using an interval with a Timer, we can
instead wait for that duration between invocations of the method and
impose a separate, hardcoded timeout to it.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>