Ensure counter vectors start with a zero value
On application restart, counter vectors such as the apply count will
only start being exported on the first update, with a value of one. This
can cause issues with graphs missing the reset and initial change. By
fetching the vector the prometheus also initialises it and exports it
with a value of zero, if it does not already exist.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>