Enable use of multiple SSH deploy keys to clone private repos
Switch from SSH flags to SSH config file for kustomize builds. This
allows us to construct SSH config with custom Hosts that correspond to
the private keys we have.

We also replace the host part of the remote repo address with the custom
value to match the Host in SSH config and use the correct corresponding
key