Update release process
Remove explicit versions from documentation and examples, I don't think
it's valuable and it adds the burden of having to update it every time
we release.

Modify the release target to update the image tag in the manifests to
match the release version and then revert back to latest. This gives us
a commit to tag that allows you to pull manifests for a release version
but also means that using master as the ref gives you latest.