Check access before adding a resource to the list of prunable resources
The delegate service account kube-applier is using may not have the
adequate permissions required to prune all resources supported by the
cluster. In fact, it's highly likely they don't as a typical namespace
admin will not have access to things like LimitRanges or other resources
they aren't intended to manage.

We can perform a SelfSubjectRulesReview on behalf of the delegate
service account and filter the list of prunable resources using that to
arrive at a list that the service account will have permissions to
prune.