webserver: retrieve Waybills on request
The minute-long delay on the status being updated can be quite
frustrating when you're force-running things and waiting for results to
appear.

I think it's better to return the status of the waybills to the user
from Kube when they request them.

The load on the apiserver should be mitigated by the use of a caching Kube
client, which controller-runtime provides and I will implement in a
separate commit.