webserver: add panel for Waybills without a last run
A new Waybill won't appear on the UI if there's some kind of setup error (i.e a
problem fetching the delegate secret) or if it's created with
`autoApply: false` because there's no 'LastRun' status field.

I've had a few reports from users where this has created confusion, so
I've added a panel that shows 'Pending' Waybills that are yet to have
their first run.

In addition to this, I've also added events to the UI, which should give
users a decent amount of context to help them understand what's going wrong and
maybe fix it themselves without having to contact us.