runner: get latest version of Waybill before status update
If you're applying a Waybill with itself (which is what we do) then you
can run into a situation where:
1. KA retrieves the Waybill and starts a run
1. During the run, `kubectl apply` updates the Waybill
2. An error is produced when KA attempts to update the Waybill's status
   at the end of the run because 'the object has been modified' (because
   `kubectl` has changed it)

To mitigate this, get the latest version of the Waybill before updating
the status.

Also add an event when status updates fail (for other reasons) so there's
some information on the UI/in kube for the user as to why their manifests
seem to have been applied but the status hasn't been updated.